update: move IProcessorListener to its own file
